% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricFamilySample.R
\name{MetricFamilySample}
\alias{MetricFamilySample}
\title{Prometheus Raw Sample Object}
\description{
Prometheus Raw Sample Object

Prometheus Raw Sample Object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetricFamilySample$new()}}
\item \href{#method-getName}{\code{MetricFamilySample$getName()}}
\item \href{#method-setName}{\code{MetricFamilySample$setName()}}
\item \href{#method-getLabelNames}{\code{MetricFamilySample$getLabelNames()}}
\item \href{#method-setLabelNames}{\code{MetricFamilySample$setLabelNames()}}
\item \href{#method-getType}{\code{MetricFamilySample$getType()}}
\item \href{#method-setType}{\code{MetricFamilySample$setType()}}
\item \href{#method-getHelp}{\code{MetricFamilySample$getHelp()}}
\item \href{#method-setHelp}{\code{MetricFamilySample$setHelp()}}
\item \href{#method-getSamples}{\code{MetricFamilySample$getSamples()}}
\item \href{#method-setSamples}{\code{MetricFamilySample$setSamples()}}
\item \href{#method-clone}{\code{MetricFamilySample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setName"></a>}}
\if{latex}{\out{\hypertarget{method-setName}{}}}
\subsection{Method \code{setName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$setName(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabelNames"></a>}}
\if{latex}{\out{\hypertarget{method-getLabelNames}{}}}
\subsection{Method \code{getLabelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$getLabelNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLabelNames"></a>}}
\if{latex}{\out{\hypertarget{method-setLabelNames}{}}}
\subsection{Method \code{setLabelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$setLabelNames(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getType"></a>}}
\if{latex}{\out{\hypertarget{method-getType}{}}}
\subsection{Method \code{getType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$getType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setType"></a>}}
\if{latex}{\out{\hypertarget{method-setType}{}}}
\subsection{Method \code{setType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$setType(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHelp"></a>}}
\if{latex}{\out{\hypertarget{method-getHelp}{}}}
\subsection{Method \code{getHelp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$getHelp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setHelp"></a>}}
\if{latex}{\out{\hypertarget{method-setHelp}{}}}
\subsection{Method \code{setHelp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$setHelp(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSamples"></a>}}
\if{latex}{\out{\hypertarget{method-getSamples}{}}}
\subsection{Method \code{getSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$getSamples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSamples"></a>}}
\if{latex}{\out{\hypertarget{method-setSamples}{}}}
\subsection{Method \code{setSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$setSamples(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricFamilySample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
