% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheusMemoryAdapter.R
\name{PrometheusMemoryAdapter}
\alias{PrometheusMemoryAdapter}
\title{Memory adapter}
\description{
Internal memory structure for storing Prometheus metrics
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrometheusMemoryAdapter$new()}}
\item \href{#method-collect}{\code{PrometheusMemoryAdapter$collect()}}
\item \href{#method-updateGauge}{\code{PrometheusMemoryAdapter$updateGauge()}}
\item \href{#method-updateCounter}{\code{PrometheusMemoryAdapter$updateCounter()}}
\item \href{#method-clone}{\code{PrometheusMemoryAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusMemoryAdapter$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
instance of PrometheusMemoryAdapter
Gather up all known metric samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collect"></a>}}
\if{latex}{\out{\hypertarget{method-collect}{}}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusMemoryAdapter$collect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of metric samples
Upsert a given gauge in memory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateGauge"></a>}}
\if{latex}{\out{\hypertarget{method-updateGauge}{}}}
\subsection{Method \code{updateGauge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusMemoryAdapter$updateGauge(input_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_list}}{list of values needed to update a gauge}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PrometheusGauge being updated
Upsert a given counter in memory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateCounter"></a>}}
\if{latex}{\out{\hypertarget{method-updateCounter}{}}}
\subsection{Method \code{updateCounter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusMemoryAdapter$updateCounter(input_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_list}}{list of values needed to update a counter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PrometheusCounter being updated
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusMemoryAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
