% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheusRenderMetrics.R
\name{PrometheusRenderMetrics}
\alias{PrometheusRenderMetrics}
\title{Output for Prometheus to parse}
\description{
Generally call this as a /metrics end point in plumber
}
\examples{

## ------------------------------------------------
## Method `PrometheusRenderMetrics$render`
## ------------------------------------------------

registry <<- CollectorRegistry$new()
renderer <- PrometheusRenderMetrics$new()
out <- renderer$render(registry$getMetricFamilySamples())
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrometheusRenderMetrics$new()}}
\item \href{#method-render}{\code{PrometheusRenderMetrics$render()}}
\item \href{#method-clone}{\code{PrometheusRenderMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusRenderMetrics$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
instance of PrometheusRenderMetrics
Render text in the Prometheus format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusRenderMetrics$render(metrics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metrics}}{list of MetricFamilySample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{registry <<- CollectorRegistry$new()
renderer <- PrometheusRenderMetrics$new()
out <- renderer$render(registry$getMetricFamilySamples())
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusRenderMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
