% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheusShiny.R
\name{prometheusRenderShiny}
\alias{prometheusRenderShiny}
\title{Prometheus with Shiny}
\usage{
prometheusRenderShiny(app, registry)
}
\arguments{
\item{app}{A shiny application as returned by [shiny::shinyApp()].}

\item{registry}{A registry as returned by [CollectorRegistry].}
}
\description{
Render the metrics and serve the `/metrics` 
endpoint in shiny.semantic
}
\examples{
library(shiny)

registry <<- CollectorRegistry$new()
registry$registerGauge(
 name = 'test',
 help = 'some_gauge',
 namespace = "my_space"
)

ui <- fluidPage(
 actionButton("click", "Click")
)

server <- function(input, output){
 gauge <- registry$getGauge(name = 'test', namespace = "my_space")

 observeEvent(input$click, {
   gauge$inc()
 })
}


app <- shinyApp(ui, server)

if(interactive())
 prometheusRenderShiny(app, registry)

}
