% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectorRegistry.R
\name{CollectorRegistry}
\alias{CollectorRegistry}
\title{Prometheus collector registry}
\description{
Prometheus collector registry

Prometheus collector registry
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CollectorRegistry$new()}}
\item \href{#method-getMetricFamilySamples}{\code{CollectorRegistry$getMetricFamilySamples()}}
\item \href{#method-registerGauge}{\code{CollectorRegistry$registerGauge()}}
\item \href{#method-getGauge}{\code{CollectorRegistry$getGauge()}}
\item \href{#method-registerCounter}{\code{CollectorRegistry$registerCounter()}}
\item \href{#method-getCounter}{\code{CollectorRegistry$getCounter()}}
\item \href{#method-clone}{\code{CollectorRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectorRegistry$new(storage_adapter = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetricFamilySamples"></a>}}
\if{latex}{\out{\hypertarget{method-getMetricFamilySamples}{}}}
\subsection{Method \code{getMetricFamilySamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectorRegistry$getMetricFamilySamples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerGauge"></a>}}
\if{latex}{\out{\hypertarget{method-registerGauge}{}}}
\subsection{Method \code{registerGauge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectorRegistry$registerGauge(name, help, namespace = NULL, labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGauge"></a>}}
\if{latex}{\out{\hypertarget{method-getGauge}{}}}
\subsection{Method \code{getGauge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectorRegistry$getGauge(name, namespace = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerCounter"></a>}}
\if{latex}{\out{\hypertarget{method-registerCounter}{}}}
\subsection{Method \code{registerCounter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectorRegistry$registerCounter(
  name,
  help,
  namespace = NULL,
  labels = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCounter"></a>}}
\if{latex}{\out{\hypertarget{method-getCounter}{}}}
\subsection{Method \code{getCounter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectorRegistry$getCounter(name, namespace = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectorRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
