% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheusCounter.R
\name{PrometheusCounter}
\alias{PrometheusCounter}
\title{Prometheus Counter implementation}
\description{
Prometheus Counter implementation

Prometheus Counter implementation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrometheusCounter$new()}}
\item \href{#method-getType}{\code{PrometheusCounter$getType()}}
\item \href{#method-getName}{\code{PrometheusCounter$getName()}}
\item \href{#method-getHelp}{\code{PrometheusCounter$getHelp()}}
\item \href{#method-getLabelNames}{\code{PrometheusCounter$getLabelNames()}}
\item \href{#method-getKey}{\code{PrometheusCounter$getKey()}}
\item \href{#method-set}{\code{PrometheusCounter$set()}}
\item \href{#method-incBy}{\code{PrometheusCounter$incBy()}}
\item \href{#method-inc}{\code{PrometheusCounter$inc()}}
\item \href{#method-clone}{\code{PrometheusCounter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$new(
  storage_adapter,
  name,
  help,
  namespace = NULL,
  label_names = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getType"></a>}}
\if{latex}{\out{\hypertarget{method-getType}{}}}
\subsection{Method \code{getType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$getType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHelp"></a>}}
\if{latex}{\out{\hypertarget{method-getHelp}{}}}
\subsection{Method \code{getHelp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$getHelp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabelNames"></a>}}
\if{latex}{\out{\hypertarget{method-getLabelNames}{}}}
\subsection{Method \code{getLabelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$getLabelNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getKey"></a>}}
\if{latex}{\out{\hypertarget{method-getKey}{}}}
\subsection{Method \code{getKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$getKey()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$set(value, labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-incBy"></a>}}
\if{latex}{\out{\hypertarget{method-incBy}{}}}
\subsection{Method \code{incBy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$incBy(value, labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\if{latex}{\out{\hypertarget{method-inc}{}}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$inc(labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusCounter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
