% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheusGauge.R
\name{PrometheusGauge}
\alias{PrometheusGauge}
\title{Prometheus collector registry}
\description{
Prometheus collector registry

Prometheus collector registry
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrometheusGauge$new()}}
\item \href{#method-getType}{\code{PrometheusGauge$getType()}}
\item \href{#method-getName}{\code{PrometheusGauge$getName()}}
\item \href{#method-getHelp}{\code{PrometheusGauge$getHelp()}}
\item \href{#method-getLabelNames}{\code{PrometheusGauge$getLabelNames()}}
\item \href{#method-getKey}{\code{PrometheusGauge$getKey()}}
\item \href{#method-set}{\code{PrometheusGauge$set()}}
\item \href{#method-incBy}{\code{PrometheusGauge$incBy()}}
\item \href{#method-inc}{\code{PrometheusGauge$inc()}}
\item \href{#method-decBy}{\code{PrometheusGauge$decBy()}}
\item \href{#method-dec}{\code{PrometheusGauge$dec()}}
\item \href{#method-clone}{\code{PrometheusGauge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$new(
  storage_adapter,
  name,
  help,
  namespace = NULL,
  label_names = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getType"></a>}}
\if{latex}{\out{\hypertarget{method-getType}{}}}
\subsection{Method \code{getType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$getType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHelp"></a>}}
\if{latex}{\out{\hypertarget{method-getHelp}{}}}
\subsection{Method \code{getHelp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$getHelp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabelNames"></a>}}
\if{latex}{\out{\hypertarget{method-getLabelNames}{}}}
\subsection{Method \code{getLabelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$getLabelNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getKey"></a>}}
\if{latex}{\out{\hypertarget{method-getKey}{}}}
\subsection{Method \code{getKey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$getKey()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$set(value, labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-incBy"></a>}}
\if{latex}{\out{\hypertarget{method-incBy}{}}}
\subsection{Method \code{incBy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$incBy(value, labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\if{latex}{\out{\hypertarget{method-inc}{}}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$inc(labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decBy"></a>}}
\if{latex}{\out{\hypertarget{method-decBy}{}}}
\subsection{Method \code{decBy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$decBy(value, labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dec"></a>}}
\if{latex}{\out{\hypertarget{method-dec}{}}}
\subsection{Method \code{dec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$dec(labels = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusGauge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
