% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrometheusSample.R
\name{PrometheusSample}
\alias{PrometheusSample}
\title{Prometheus Raw Sample Object}
\description{
Prometheus Raw Sample Object

Prometheus Raw Sample Object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PrometheusSample$new()}}
\item \href{#method-getName}{\code{PrometheusSample$getName()}}
\item \href{#method-setName}{\code{PrometheusSample$setName()}}
\item \href{#method-getLabelNames}{\code{PrometheusSample$getLabelNames()}}
\item \href{#method-setLabelNames}{\code{PrometheusSample$setLabelNames()}}
\item \href{#method-getLabelValues}{\code{PrometheusSample$getLabelValues()}}
\item \href{#method-setLabelValues}{\code{PrometheusSample$setLabelValues()}}
\item \href{#method-getValue}{\code{PrometheusSample$getValue()}}
\item \href{#method-setValue}{\code{PrometheusSample$setValue()}}
\item \href{#method-hasLabelNames}{\code{PrometheusSample$hasLabelNames()}}
\item \href{#method-clone}{\code{PrometheusSample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$new(name, value, label_names = list(), label_values = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setName"></a>}}
\if{latex}{\out{\hypertarget{method-setName}{}}}
\subsection{Method \code{setName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$setName(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabelNames"></a>}}
\if{latex}{\out{\hypertarget{method-getLabelNames}{}}}
\subsection{Method \code{getLabelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$getLabelNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLabelNames"></a>}}
\if{latex}{\out{\hypertarget{method-setLabelNames}{}}}
\subsection{Method \code{setLabelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$setLabelNames(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabelValues"></a>}}
\if{latex}{\out{\hypertarget{method-getLabelValues}{}}}
\subsection{Method \code{getLabelValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$getLabelValues()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setLabelValues"></a>}}
\if{latex}{\out{\hypertarget{method-setLabelValues}{}}}
\subsection{Method \code{setLabelValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$setLabelValues(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-getValue}{}}}
\subsection{Method \code{getValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$getValue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setValue"></a>}}
\if{latex}{\out{\hypertarget{method-setValue}{}}}
\subsection{Method \code{setValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$setValue(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasLabelNames"></a>}}
\if{latex}{\out{\hypertarget{method-hasLabelNames}{}}}
\subsection{Method \code{hasLabelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$hasLabelNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PrometheusSample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
